backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
click==8.1.8
exceptiongroup==1.3.0
fastapi==0.116.1
h11==0.16.0
httptools==0.6.4
idna==3.10
paho-mqtt>=2.0.0
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==1.1.1
PyYAML==6.0.2
sniffio==1.3.1
sqlalchemy>=2.0.0
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
